fix dep cycle
