+++ src/api/providers/kilocode-openrouter.ts
@@ -3,6 +3,7 @@ import { OpenRouterHandler } from "./openrouter"
 import { getModelParams } from "../transform/model-params"
 import { getModels } from "./fetchers/modelCache"
 import { DEEP_SEEK_DEFAULT_TEMPERATURE } from "./constants"
+import { getKiloBaseUriFromToken } from "../../utils/kilocode-token"
 
 /**
  * A custom OpenRouter handler that overrides the getModel function
@@ -77,16 +78,4 @@ export class KilocodeOpenrouterHandler extends OpenRouterHandler {
 
 function getKiloBaseUri(options: ApiHandlerOptions) {
 	return getKiloBaseUriFromToken(options.kilocodeToken ?? "")
-}
-
-export function getKiloBaseUriFromToken(kilocodeToken: string) {
-	try {
-		const payload_string = kilocodeToken.split(".")[1]
-		const payload = JSON.parse(Buffer.from(payload_string, "base64").toString())
-		//note: this is UNTRUSTED, so we need to make sure we're OK with this being manipulated by an attacker; e.g. we should not read uri's from the JWT directly.
-		if (payload.env === "development") return "http://localhost:3000"
-	} catch (_error) {
-		console.warn("Failed to get base URL from Kilo Code token")
-	}
-	return "https://kilocode.ai"
 }
